requirements.txt
fastapi[all] #includes websocket, uvicorn
earthaccess
fiona==1.8.22
xarray
geopandas
python-dotenv
h5netcdf